modify to use Ceres solver

+++ lib/matrix/vector.h
@@ -42,7 +42,8 @@ typedef struct {
 
 /* 3D vector of ints */
 typedef struct {
-    int32_t p[3];
+    //int32_t p[3];
+	unsigned long p[3];
 } i32_v3_t;
 
 /* 3D vector of ints */
